Fix plotting a DataFrame without 0 in the index
This commit prevents the KeyError raised when DataFrame.plot() is called
with xerr or yerr being a Series or DataFrame whose index doesn't include 0.
The error comes from matplotlib code which tries to access xerr[0] or yerr[0],
so to solve the problem, we convert xerr and yerr from Pandas objects to
NumPy ndarrays before sending them through to matplotlib. This is
a different instance of the same type of problem in Github issues #4493
and #6127 (and perhaps others).